changed int to uintptr_t for all pointer type casts

+++ code/vision/trunk/subarchitectures/vision.sa/src/c++/vision/components/SOIFilter/graph.cpp
@@ -86,7 +86,7 @@ void Graph::add_edge(node_id from, node_id to, captype cap, captype rev_cap)
 		arc_for_block *next = arc_for_block_first;
 		char *ptr = new char[sizeof(arc_for_block)+1];
 		if (!ptr) { if (error_function) (*error_function)("Not enough memory!"); exit(1); }
-		if ((int)ptr & 1) arc_for_block_first = (arc_for_block *) (ptr + 1);
+		if ((uintptr_t)ptr & 1) arc_for_block_first = (arc_for_block *) (ptr + 1);
 		else              arc_for_block_first = (arc_for_block *) ptr;
 		arc_for_block_first -> start = ptr;
 		arc_for_block_first -> current = & ( arc_for_block_first -> arcs_for[0] );
@@ -98,7 +98,7 @@ void Graph::add_edge(node_id from, node_id to, captype cap, captype rev_cap)
 		arc_rev_block *next = arc_rev_block_first;
 		char *ptr = new char[sizeof(arc_rev_block)+1];
 		if (!ptr) { if (error_function) (*error_function)("Not enough memory!"); exit(1); }
-		if ((int)ptr & 1) arc_rev_block_first = (arc_rev_block *) (ptr + 1);
+		if ((uintptr_t)ptr & 1) arc_rev_block_first = (arc_rev_block *) (ptr + 1);
 		else              arc_rev_block_first = (arc_rev_block *) ptr;
 		arc_rev_block_first -> start = ptr;
 		arc_rev_block_first -> current = & ( arc_rev_block_first -> arcs_rev[0] );
@@ -109,14 +109,14 @@ void Graph::add_edge(node_id from, node_id to, captype cap, captype rev_cap)
 	a_rev = arc_rev_block_first -> current ++;
 
 	a_rev -> sister = (arc_forward *) from;
-	a_for -> shift  = (int) to;
+	a_for -> shift  = (uintptr_t) to;
 	a_for -> r_cap = cap;
 	a_for -> r_rev_cap = rev_cap;
 
 	((node *)from) -> first_out =
-		(arc_forward *) ((int)(((node *)from) -> first_out) + 1);
+		(arc_forward *) ((uintptr_t)(((node *)from) -> first_out) + 1);
 	((node *)to) -> first_in =
-		(arc_reverse *) ((int)(((node *)to) -> first_in) + 1);
+		(arc_reverse *) ((uintptr_t)(((node *)to) -> first_in) + 1);
 }
 
 void Graph::set_tweights(node_id i, captype cap_source, captype cap_sink)
@@ -155,7 +155,7 @@ void Graph::prepare_graph()
 	arc_reverse *a_rev, *a_rev_scan, a_rev_tmp;
 	node_block *nb;
 	bool for_flag = false, rev_flag = false;
-	int k;
+	uintptr_t k;
 
 	if (!arc_rev_block_first)
 	{
@@ -180,7 +180,7 @@ void Graph::prepare_graph()
 		for (i=&nb->nodes[0]; i<nb->current; i++)
 		{
 			/* outgoing arcs */
-			k = (int) i -> first_out;
+			k = (uintptr_t) i -> first_out;
 			if (a_for + k > &ab_for->arcs_for[ARC_BLOCK_SIZE])
 			{
 				if (k > ARC_BLOCK_SIZE) { if (error_function) (*error_function)("# of arcs per node exceeds block size!"); exit(1); }
@@ -191,7 +191,7 @@ void Graph::prepare_graph()
 					arc_for_block *next = arc_for_block_first;
 					char *ptr = new char[sizeof(arc_for_block)+1];
 					if (!ptr) { if (error_function) (*error_function)("Not enough memory!"); exit(1); }
-					if ((int)ptr & 1) arc_for_block_first = (arc_for_block *) (ptr + 1);
+					if ((uintptr_t)ptr & 1) arc_for_block_first = (arc_for_block *) (ptr + 1);
 					else              arc_for_block_first = (arc_for_block *) ptr;
 					arc_for_block_first -> start = ptr;
 					arc_for_block_first -> current = & ( arc_for_block_first -> arcs_for[0] );
@@ -213,7 +213,7 @@ void Graph::prepare_graph()
 			ab_for -> last_node = i;
 
 			/* incoming arcs */
-			k = (int) i -> first_in;
+			k = (uintptr_t) i -> first_in;
 			if (a_rev + k > &ab_rev->arcs_rev[ARC_BLOCK_SIZE])
 			{
 				if (k > ARC_BLOCK_SIZE) { if (error_function) (*error_function)("# of arcs per node exceeds block size!"); exit(1); }
@@ -224,7 +224,7 @@ void Graph::prepare_graph()
 					arc_rev_block *next = arc_rev_block_first;
 					char *ptr = new char[sizeof(arc_rev_block)+1];
 					if (!ptr) { if (error_function) (*error_function)("Not enough memory!"); exit(1); }
-					if ((int)ptr & 1) arc_rev_block_first = (arc_rev_block *) (ptr + 1);
+					if ((uintptr_t)ptr & 1) arc_rev_block_first = (arc_rev_block *) (ptr + 1);
 					else              arc_rev_block_first = (arc_rev_block *) ptr;
 					arc_rev_block_first -> start = ptr;
 					arc_rev_block_first -> current = & ( arc_rev_block_first -> arcs_rev[0] );
@@ -303,7 +303,7 @@ void Graph::prepare_graph()
 		ab_for -> current -> shift     = a_for -> shift;
 		ab_for -> current -> r_cap     = a_for -> r_cap;
 		ab_for -> current -> r_rev_cap = a_for -> r_rev_cap;
-		a_for -> shift = (int) (ab_for -> current + 1);
+		a_for -> shift = (uintptr_t) (ab_for -> current + 1);
 		i -> first_out = (arc_forward *) (((char *)a_for) - 1);
 	}
